Adding untested code for creating new files and running terminal commands.
Parsing, Resolving and application included.
Prompt is not updated yet. I'm thinking I should include a configuration to add prompts for the commands and creating new files or not, as this will further distract the model. I can have two commands for the demo for best results??
Probably will not have enough time